Convert antares output to h5 file

Convert antares output to h5 file

writeAntaresH5(
  path = NULL,
  timeSteps = c("hourly", "daily", "weekly", "monthly", "annual"),
  opts = simOptions(),
  writeMcAll = TRUE,
  compress = 1,
  misc = FALSE,
  thermalAvailabilities = FALSE,
  hydroStorage = FALSE,
  hydroStorageMaxPower = FALSE,
  reserve = FALSE,
  linkCapacity = FALSE,
  mustRun = FALSE,
  thermalModulation = FALSE,
  allData = FALSE,
  writeAllSimulations = FALSE,
  nbCores = 4,
  removeVirtualAreas = FALSE,
  storageFlexibility = NULL,
  production = NULL,
  reassignCosts = FALSE,
  newCols = TRUE,
  overwrite = FALSE,
  supressMessages = FALSE
)

Arguments

path
character folder where h5 file will be write (default NULL)

timeSteps
character timeSteps

opts
list of simulation parameters returned by the function setSimulationPath. Defaut to antaresRead::simOptions()

writeMcAll
boolean write mc-all

compress
numeric compress level

misc
boolean see readAntares

thermalAvailabilities
boolean see readAntares

hydroStorage
boolean see readAntares

hydroStorageMaxPower
boolean see readAntares

reserve
boolean see readAntares

linkCapacity
boolean see readAntares

mustRun
boolean see readAntares

thermalModulation
boolean see readAntares

allData
boolean add all data with a single call (writeMcAll, misc, thermalAvailabilities, hydroStorage, hydroStorageMaxPower reserve, linkCapacity, mustRun, thermalModulation).

writeAllSimulations
boolean, write all simulations of your antares study.

nbCores
numeric, number of cores to use, only used if writeAllSimulations is TRUE

removeVirtualAreas
boolean, remove virtual areas, see removeVirtualAreas

storageFlexibility
character or list, see removeVirtualAreas

production
character or list, see removeVirtualAreas

reassignCosts
boolean or list, see removeVirtualAreas

newCols
boolean or list, see removeVirtualAreas

overwrite
boolean or list, overwrite old file

supressMessages
boolean, supress messages from readAntares and removeVirtualAreas

Examples

if (FALSE) {
# Write simulation one by one
setSimulationPath("C:/Users/MyUser/Mystudy", 1)
writeAntaresH5(path="PATH_TO_YOUR_STUDY")

# Write all simulations
setSimulationPath("C:/Users/MyUser/Mystudy")
writeAntaresH5(path="PATH_TO_YOUR_STUDY", writeAllSimulations = TRUE)

# Choose timestep to write
setSimulationPath("C:/Users/MyUser/Mystudy", 1)
writeAntaresH5(path="PATH_TO_YOUR_STUDY", timeSteps = "hourly")

# Write with additionnal information
writeAntaresH5(path="PATH_TO_YOUR_STUDY", timeSteps = "hourly",
   misc = TRUE, thermalAvailabilities = TRUE,
   hydroStorage = TRUE, hydroStorageMaxPower = TRUE, reserve = TRUE,
   linkCapacity = TRUE, mustRun = TRUE, thermalModulation = TRUE)

# Write all data with a shorcut 
writeAntaresH5(path="PATH_TO_YOUR_STUDY", allData = TRUE)

#Remove virtuals areas

writeAntaresH5(path="PATH_TO_YOUR_STUDY", timeSteps = "hourly", overwrite = TRUE,
               writeMcAll = FALSE, removeVirtualAreas = TRUE,
               storageFlexibility =  "psp in-2",
               production =  NULL, reassignCosts =FALSE, newCols = TRUE)

#Remove virtuals areas more than one call
writeAntaresH5(
               path="PATH_TO_YOUR_STUDY",
               timeSteps = "hourly",
               overwrite = TRUE,
               writeMcAll = FALSE,
               removeVirtualAreas = TRUE,
               storageFlexibility = list("psp out", "psp in-2"),
               production = list(NULL, NULL),
               reassignCosts = list(TRUE, FALSE),
               newCols = list(FALSE, TRUE)
               )

}