readBindingConstraints {antaresRead}	R Documentation

Read binding constraints

Description

This function reads the binding constraints of an Antares project.

Be aware that binding constraints are read in the input files of a study. So they may have changed since a simulation has been run.

Usage

readBindingConstraints(opts = simOptions())

## S3 method for class 'bindingConstraints'
summary(object, ...)

Arguments

opts
list of simulation parameters returned by the function setSimulationPath

object
Object returned by readBindingConstraints

...
Unused

Value

readBindingConstraints returns an object of class bindingConstraints. It is a named list with one element per read constraint. Each element is itself a list with the following elements:

enabled
is the constraint enabled ?

timeStep
time step the constraint applies to

operator
type of constraint: equality, inequality on one side or both sides

coefficients
elements containing the coefficients used by the constraint

values
values used by the constraint. It contains one line per time step and three columns "less", "greater" and "equal"

The summary method returns a data.frame with one line per constraint.

Examples

## Not run: 
setSimulationPath()

constraints <- readBindingConstraints()
summary(constraints)

## End(Not run)

[Package antaresRead version 2.2.91 Index]