Import clusters description

This function reads in the input files of an antares study the characteristics of each cluster.

Be aware that clusters descriptions are read in the input files so they may have changed since a simulation has been run.

readClusterDesc(opts = simOptions())

readClusterResDesc(opts = simOptions())

readClusterSTDesc(opts = simOptions())

Arguments

opts

list of simulation parameters returned by the function setSimulationPath

Value

A data.table with one line per cluster. The columns of the data.table may change between different projects, but there will always be the following columns:

area

Name of the area containing the cluster

cluster

Name of the cluster

group

Type of cluster (gaz, nuclear, etc.)

unitcount

number of production units

nominalcapacity

production capacity of each unit

The other present columns depends on the version of antares and the options that have been set: if an option is unset for all clusters, it will not appear in the table.

By default, the function reads the cluster description of the default antares study. You can use the argument opts to specify another study.

readClusterDesc : read thermal clusters

readClusterResDesc : read renewable clusters (Antares >= V8.1)

readClusterSTDesc : read st-storage clusters (Antares >= V8.6)

If you have no clusters properties, Null data.table (0 rows and 0 cols) is returned.

Examples

if (FALSE) { # \dontrun{

# thermal
readClusterDesc()

# renewable
readClusterResDesc()

# st-storage
readClusterSTDesc()

# By default, the function reads cluster descriptions for the default study,
# but it is possible to specify another study with parameter "opts"
sim1 <- setSimulationPath()

#[... code that modifies the default antares study]

readClusterDesc(sim1)

} # }