simOptions {antaresRead}	R Documentation

Extract simulation options

Description

The function readAntares stores in its output the options used to read some data (path of the study, area list, link list, start date, etc.).

Usage

simOptions(x = NULL)

Arguments

x
object of class antaresTable or antaresData

Details

simOptions extracts these options from an object of class antaresTable or antaresOutput. It can be useful when working on multiple simulations, either to check how some object has been created or to use it in some functions like getAreas or getLinks

If the parameter of the function is NULL, it returns the default simulation options, that is the options set by setSimulationPath the last time it was run.

Value

list of options used to read the data contained in an object or the last simulation options read by setSimulationPath if x is NULL

Examples

## Not run: 
  setSimulationPath(study1)
  
  simOptions() # returns the options for study 1
  
  data <- readAntares()
  
  # Choose a different study
  setSimulationPath(study2)
  
  simOptions() # returns the options for study 2
  
  getAreas() # returns the areas of the secund study
  getAreas(opts = simOptions(data)) # returns the areas of the first study
  

## End(Not run)

[Package antaresRead version 2.2.91 Index]